requirements.txt
flask
flask-cors
flask-restful
flask-sqlalchemy
pyjwt
pymysql